pyproject.toml
[tool.poetry]
name = "trovo-bot"
version = "0.1.0"
description = ""
authors = ["Yuriy Grebenshchikov <grur.ura@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apscheduler = "3.11.0"
pyjwt = ">=2.8.0"
sqlalchemy = ">=2.0.9,<2.1.0"
aiohttp = "3.10.11"
alembic = ">=1.11.1"
bcrypt = "4.3.0"
cryptography = "44.0.1"
fastapi = ">=0.101.0"
loguru = ">=0.7.0"
numpy = ">=1.25.0"
passlib = ">=1.7.4,<1.8.0"
psycopg2-binary = ">=2.9.5,<2.10.0"
pydantic = { version = "2.11.3", extras = ["email"] }
pydantic-settings = ">=2.0.0"
python-dotenv = ">=1.0.0"
python-socketio = "5.12.1"
requests = "2.32.3"
tenacity = ">=8.1.0"
unidecode = ">=1.3.0"
uvicorn = "0.34.0"
websockets = ">=11"

[tool.poetry.dev-dependencies]
black = ">=23.0.0"

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py311']
exclude = '''
            (
              /(
                \.git
                | venv
                | .venv
              )/
            )
          '''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
